+++ rocketgram/api/__init__.py
@@ -167,6 +167,7 @@ from .sticker_set import StickerSet
 from .stop_message_live_location import StopMessageLiveLocation
 from .stop_poll import StopPoll
 from .successful_payment import SuccessfulPayment
+from .thumb_mime_type import ThumbMimeType
 from .unban_chat_member import UnbanChatMember
 from .unpin_chat_message import UnpinChatMessage
 from .update import Update